refactor(store): clean up
