+++ src/store/actions.ts
@@ -239,39 +239,3 @@ export const setConnectionMode = (connectionMode: ConnectionMode): ActionTypes =
   type: SET_CONNECTION_MODE,
   payload: { connectionMode },
 });
-
-export const actions = {
-  setOnConnect,
-  setOnConnectStart,
-  setOnConnectStop,
-  setOnConnectEnd,
-  setElements,
-  batchUpdateNodeDimensions,
-  updateNodeDimensions,
-  updateNodePos,
-  updateNodePosDiff,
-  setUserSelection,
-  updateUserSelection,
-  unsetUserSelection,
-  setSelection,
-  unsetNodesSelection,
-  resetSelectedElements,
-  setSelectedElements,
-  addSelectedElements,
-  updateTransform,
-  updateSize,
-  initD3Zoom,
-  setMinZoom,
-  setMaxZoom,
-  setTranslateExtent,
-  setConnectionPosition,
-  setConnectionNodeId,
-  setSnapToGrid,
-  setSnapGrid,
-  setInteractive,
-  setNodesDraggable,
-  setNodesConnectable,
-  setElementsSelectable,
-  setMultiSelectionActive,
-  setConnectionMode,
-};
