+++ src/store/configure-store.ts
@@ -1,7 +1,8 @@
 import { applyMiddleware, compose, createStore, Store } from 'redux';
 import thunkMiddleware from 'redux-thunk';
 
-import { ReactFlowState, reactFlowReducer } from './index';
+import { ReactFlowState } from '../types';
+import reactFlowReducer from './reducer';
 
 export default function configureStore(preloadedState: ReactFlowState): Store<ReactFlowState> {
   const composedEnhancers = compose(applyMiddleware(thunkMiddleware));
