+++ src/types/index.ts
@@ -373,3 +373,49 @@ export type InitD3ZoomPayload = {
   d3ZoomHandler: ((this: Element, event: any, d: unknown) => void) | undefined;
   transform: Transform;
 };
+
+export interface ReactFlowState {
+  width: number;
+  height: number;
+  transform: Transform;
+  elements: Elements;
+  // nodes: Computed<StoreModel, Node[]>;
+  // edges: Computed<StoreModel, Edge[]>;
+  selectedElements: Elements | null;
+  selectedNodesBbox: Rect;
+  // viewportBox: Computed<StoreModel, Rect>;
+
+  d3Zoom: ZoomBehavior<Element, unknown> | null;
+  d3Selection: D3Selection<Element, unknown, null, undefined> | null;
+  d3ZoomHandler: ((this: Element, event: any, d: unknown) => void) | undefined;
+  minZoom: number;
+  maxZoom: number;
+  translateExtent: TranslateExtent;
+
+  nodesSelectionActive: boolean;
+  selectionActive: boolean;
+
+  userSelectionRect: SelectionRect;
+
+  connectionNodeId: ElementId | null;
+  connectionHandleId: ElementId | null;
+  connectionHandleType: HandleType | null;
+  connectionPosition: XYPosition;
+  connectionMode: ConnectionMode;
+
+  snapToGrid: boolean;
+  snapGrid: SnapGrid;
+
+  nodesDraggable: boolean;
+  nodesConnectable: boolean;
+  elementsSelectable: boolean;
+
+  multiSelectionActive: boolean;
+
+  reactFlowVersion: string;
+
+  onConnect?: OnConnectFunc;
+  onConnectStart?: OnConnectStartFunc;
+  onConnectStop?: OnConnectStopFunc;
+  onConnectEnd?: OnConnectEndFunc;
+}
